refactor: replace server/ globals with create_mcp_app factory and click CLI
Restructures the MCP server from 9 module-level globals + bare main() to
a factory pattern (create_mcp_app(**overrides)) following champi-stt layout.

- api/server.py: all 70+ tools as closures over factory-local managers;
  FastMCP lifespan handles canvas cleanup; managers exposed on mcp instance
  for test injection
- cli.py: click group with serve, demo, validate, ipc cleanup/status;
  adds click>=8.0 dependency; entry point changed to champi_imgui.cli:cli
- tests (11 files): monkeypatch-globals replaced with create_mcp_app(manager=mock);
  _Server proxy handles backwards-compatible server.tool.fn() call sites
- docs/MCP_SERVER_STRUCTURE.md: org-wide reference for the pattern
- server/ deleted after migration completes
