+++ src/champi_imgui/api/__init__.py
@@ -1,0 +1,5 @@
+"""MCP server module for champi-imgui."""
+
+from champi_imgui.api.server import create_mcp_app
+
+__all__ = ["create_mcp_app"]
