--- src/champi_imgui/server/__init__.py
@@ -1,5 +1,0 @@
-"""MCP server module for champi-imgui."""
-
-from champi_imgui.server.main import mcp
-
-__all__ = ["mcp"]
