+++ tests/test_get_canvas_info.py
@@ -7,23 +7,54 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-import champi_imgui.server.main as server
+import champi_imgui.api.server as _srv_mod
+from champi_imgui.api.server import create_mcp_app
 from champi_imgui.core.canvas import Canvas, CanvasManager
 from champi_imgui.widgets.drawing import DrawingWidget
+
+_DEFAULT_MCP = None
+
+
+class _Server:
+    """Proxy that exposes MCP tools and managers from a factory instance.
+
+    Falls back to a lazy default instance for parametrize collection.
+    """
+
+    _mcp = None
+
+    def __getattr__(self, name: str):
+        global _DEFAULT_MCP
+        if hasattr(_srv_mod, name):
+            return getattr(_srv_mod, name)
+        mcp = type(self)._mcp
+        if mcp is None:
+            if _DEFAULT_MCP is None:
+                _DEFAULT_MCP = create_mcp_app()
+            mcp = _DEFAULT_MCP
+        if hasattr(mcp, name):
+            return getattr(mcp, name)
+        if hasattr(mcp, f"_{name}"):
+            return getattr(mcp, f"_{name}")
+        tools = mcp._tool_manager._tools
+        if name in tools:
+            return tools[name]
+        raise AttributeError(f"_Server has no attribute '{name}'")
+
+
+server = _Server()
 
 
 @pytest.fixture(autouse=True)
 def fresh_canvas_manager(monkeypatch):
-    """Swap module-level canvas_manager with a clean instance per test."""
     manager = CanvasManager()
     monkeypatch.setattr(manager, "ensure_canvas_running", lambda canvas_id: True)
-    original = server.canvas_manager
-    server.canvas_manager = manager
+    _Server._mcp = create_mcp_app(canvas_manager=manager)
     yield manager
     for canvas in list(manager.canvases.values()):
         with contextlib.suppress(Exception):
             canvas.shm_manager.cleanup()
-    server.canvas_manager = original
+    _Server._mcp = None
 
 
 @pytest.fixture()
