+++ tests/test_system_state.py
@@ -1,9 +1,9 @@
-"""Tests for to_diagnostics() methods and system-state MCP tools (issues #135–#140)."""
+"""Tests for to_diagnostics() methods and system-state MCP tools (issues #135-#140)."""
 
 from unittest.mock import MagicMock, patch
 
-
 # ─────────────────────────── AnimationManager ────────────────────────────────
+
 
 def test_animation_manager_diagnostics_empty():
     from champi_imgui.extensions.animation import AnimationManager
@@ -16,7 +16,11 @@ def test_animation_manager_diagnostics_empty():
 
 
 def test_animation_manager_diagnostics_enum_values():
-    from champi_imgui.extensions.animation import AnimationManager, AnimationState, EasingFunction
+    from champi_imgui.extensions.animation import (
+        AnimationManager,
+        AnimationState,
+        EasingFunction,
+    )
 
     mgr = AnimationManager()
     mgr.create("fade", 0.0, 1.0, 0.5, easing=EasingFunction.LINEAR)
@@ -32,6 +36,7 @@ def test_animation_manager_diagnostics_enum_values():
 
 # ─────────────────────────── NotificationManager ─────────────────────────────
 
+
 def test_notification_manager_diagnostics_empty():
     from champi_imgui.extensions.notification import NotificationManager
 
@@ -42,7 +47,10 @@ def test_notification_manager_diagnostics_empty():
 
 
 def test_notification_manager_diagnostics_with_entry():
-    from champi_imgui.extensions.notification import NotificationManager, NotificationType
+    from champi_imgui.extensions.notification import (
+        NotificationManager,
+        NotificationType,
+    )
 
     mgr = NotificationManager()
     with patch("champi_imgui.extensions.notification.imgui") as mock_imgui:
@@ -56,6 +64,7 @@ def test_notification_manager_diagnostics_with_entry():
 
 
 # ──────────────────────────── LayoutManager ──────────────────────────────────
+
 
 def test_layout_manager_diagnostics_defaults():
     from champi_imgui.layout.manager import LayoutManager, LayoutMode
@@ -80,6 +89,7 @@ def test_layout_manager_diagnostics_after_mode_change():
 
 
 # ──────────────────────────── DataStore ──────────────────────────────────────
+
 
 def test_data_store_diagnostics_empty():
     from champi_imgui.core.binding import DataStore
@@ -107,6 +117,7 @@ def test_data_store_diagnostics_with_data():
 
 # ─────────────────────────── BindingManager ──────────────────────────────────
 
+
 def test_binding_manager_diagnostics_empty():
     from champi_imgui.core.binding import BindingManager, DataStore
 
@@ -128,6 +139,7 @@ def test_binding_manager_diagnostics_with_binding():
 
 
 # ──────────────────────────── EventQueue ─────────────────────────────────────
+
 
 def test_event_queue_diagnostics_empty():
     from champi_imgui.core.events import EventQueue
@@ -151,6 +163,7 @@ def test_event_queue_diagnostics_with_subscription():
 
 # ─────────────────────────── TemplateManager ─────────────────────────────────
 
+
 def test_template_manager_diagnostics_empty(tmp_path):
     from champi_imgui.core.serialization import TemplateManager
 
@@ -162,8 +175,9 @@ def test_template_manager_diagnostics_empty(tmp_path):
 
 # ─────────────────────────── ThemeManager ────────────────────────────────────
 
+
 def test_theme_manager_diagnostics_no_theme():
-    from champi_imgui.themes.manager import Theme, ThemeManager
+    from champi_imgui.themes.manager import ThemeManager
 
     mgr = ThemeManager()
     d = mgr.to_diagnostics()
@@ -184,32 +198,69 @@ def test_theme_manager_diagnostics_after_register():
 
 # ──────────────────────── get_system_state tool ──────────────────────────────
 
-def _patch_managers(monkeypatch):
-    """Return a dict of minimal mock managers for server.main globals."""
-    from champi_imgui.server import main as m
+
+def _make_mcp_with_mock_managers():
+    """Build a minimal create_mcp_app call with all mock managers."""
+    from champi_imgui.api.server import create_mcp_app
 
     mock_canvas_mgr = MagicMock()
     mock_canvas_mgr.list_canvases.return_value = []
 
-    monkeypatch.setattr(m, "canvas_manager", mock_canvas_mgr)
-    monkeypatch.setattr(m, "animation_manager", MagicMock(**{"to_diagnostics.return_value": {"active_count": 0, "total_count": 0, "animations": []}}))
-    monkeypatch.setattr(m, "data_store", MagicMock(**{"to_diagnostics.return_value": {"key_count": 0, "signal_path_count": 0, "signal_paths": []}}))
-    monkeypatch.setattr(m, "binding_manager", MagicMock(**{"to_diagnostics.return_value": {"binding_count": 0, "paths": []}}))
-    monkeypatch.setattr(m, "layout_manager", MagicMock(**{"to_diagnostics.return_value": {"mode": "free", "spacing": 5.0, "grid_columns": 3}}))
-    monkeypatch.setattr(m, "theme_manager", MagicMock(**{"to_diagnostics.return_value": {"current": None, "available": []}}))
-    monkeypatch.setattr(m, "template_manager", MagicMock(**{"to_diagnostics.return_value": {"saved": [], "cached": []}}))
-    monkeypatch.setattr(m, "notification_manager", MagicMock(**{"to_diagnostics.return_value": {"pending_count": 0, "notifications": []}}))
-    monkeypatch.setattr(m, "event_queue", MagicMock(**{"to_diagnostics.return_value": {"pending_count": 0, "subscription_count": 0}}))
+    return create_mcp_app(
+        canvas_manager=mock_canvas_mgr,
+        animation_manager=MagicMock(
+            **{
+                "to_diagnostics.return_value": {
+                    "active_count": 0,
+                    "total_count": 0,
+                    "animations": [],
+                }
+            }
+        ),
+        data_store=MagicMock(
+            **{
+                "to_diagnostics.return_value": {
+                    "key_count": 0,
+                    "signal_path_count": 0,
+                    "signal_paths": [],
+                }
+            }
+        ),
+        binding_manager=MagicMock(
+            **{"to_diagnostics.return_value": {"binding_count": 0, "paths": []}}
+        ),
+        layout_manager=MagicMock(
+            **{
+                "to_diagnostics.return_value": {
+                    "mode": "free",
+                    "spacing": 5.0,
+                    "grid_columns": 3,
+                }
+            }
+        ),
+        theme_manager=MagicMock(
+            **{"to_diagnostics.return_value": {"current": None, "available": []}}
+        ),
+        template_manager=MagicMock(
+            **{"to_diagnostics.return_value": {"saved": [], "cached": []}}
+        ),
+        notification_manager=MagicMock(
+            **{"to_diagnostics.return_value": {"pending_count": 0, "notifications": []}}
+        ),
+        event_queue=MagicMock(
+            **{
+                "to_diagnostics.return_value": {
+                    "pending_count": 0,
+                    "subscription_count": 0,
+                }
+            }
+        ),
+    )
 
-    return mock_canvas_mgr
 
-
-def test_get_system_state_success(monkeypatch):
-    from champi_imgui.server import main as m
-
-    _patch_managers(monkeypatch)
-
-    result = m.get_system_state.fn()
+def test_get_system_state_success():
+    mcp = _make_mcp_with_mock_managers()
+    result = mcp._tool_manager._tools["get_system_state"].fn()
     assert result["success"] is True
     data = result["data"]
     assert "version" in data
@@ -224,35 +275,39 @@ def test_get_system_state_success(monkeypatch):
     assert "widget_updated" in data["signals"]
 
 
-def test_get_system_state_managers_keys(monkeypatch):
-    from champi_imgui.server import main as m
-
-    _patch_managers(monkeypatch)
-
-    result = m.get_system_state.fn()
+def test_get_system_state_managers_keys():
+    mcp = _make_mcp_with_mock_managers()
+    result = mcp._tool_manager._tools["get_system_state"].fn()
     managers = result["data"]["managers"]
     assert set(managers.keys()) == {
-        "animations", "data_store", "bindings", "layout",
-        "themes", "templates", "notifications", "events",
+        "animations",
+        "data_store",
+        "bindings",
+        "layout",
+        "themes",
+        "templates",
+        "notifications",
+        "events",
     }
 
 
 # ────────────────────────── get_canvas_diagnostics ───────────────────────────
 
-def test_get_canvas_diagnostics_not_found(monkeypatch):
-    from champi_imgui.server import main as m
+
+def test_get_canvas_diagnostics_not_found():
+    from champi_imgui.api.server import create_mcp_app
 
     mock_mgr = MagicMock()
     mock_mgr.get_canvas.return_value = None
-    monkeypatch.setattr(m, "canvas_manager", mock_mgr)
+    mcp = create_mcp_app(canvas_manager=mock_mgr)
 
-    result = m.get_canvas_diagnostics.fn("missing_canvas")
+    result = mcp._tool_manager._tools["get_canvas_diagnostics"].fn("missing_canvas")
     assert result["success"] is False
     assert "not found" in result["error"]
 
 
-def test_get_canvas_diagnostics_healthy_canvas(monkeypatch):
-    from champi_imgui.server import main as m
+def test_get_canvas_diagnostics_healthy_canvas():
+    from champi_imgui.api.server import create_mcp_app
 
     mock_widget = MagicMock()
     mock_widget.__class__.__name__ = "ButtonWidget"
@@ -273,9 +328,9 @@ def test_get_canvas_diagnostics_healthy_canvas(monkeypatch):
 
     mock_mgr = MagicMock()
     mock_mgr.get_canvas.return_value = mock_canvas
-    monkeypatch.setattr(m, "canvas_manager", mock_mgr)
+    mcp = create_mcp_app(canvas_manager=mock_mgr)
 
-    result = m.get_canvas_diagnostics.fn("canvas1")
+    result = mcp._tool_manager._tools["get_canvas_diagnostics"].fn("canvas1")
     assert result["success"] is True
     data = result["data"]
     assert data["canvas_id"] == "canvas1"
